Refactored routing and requests to backend #16
- Unified endpoints naming in url
- Made base backend url a constant with ES6 string formatting
