+++ .eslintrc.json
@@ -7,6 +7,7 @@
         "comma-dangle": "off",
         "camelcase": "error",
         "space-before-blocks": "off",
-        "space-before-function-paren":"off"
+        "space-before-function-paren":"off",
+        "template-curly-spacing":"off"
     }
 }
